Progress notifikasi: buat halaman notifikasi dan notif acara
