+++ lib/ui/notifikasi/notifikasi.dart
@@ -1,0 +1,33 @@
+import 'package:boilerplate/ui/notifikasi/notif_acara.dart';
+import 'package:flutter/material.dart';
+
+class Notifikasi extends StatelessWidget {
+  const Notifikasi({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return DefaultTabController(
+      length: 3,
+      child: Scaffold(
+        appBar: AppBar(
+          backgroundColor: Colors.green,
+          bottom: TabBar(
+            tabs: [
+              Tab(text: 'Lapor Kotor'),
+              Tab(text: 'Lapor Acara'),
+              Tab(text: 'Feedback Tugas'),
+            ],
+          ),
+          title: Text('Tabs Demo'),
+        ),
+        body: TabBarView(
+          children: [
+            Padding(padding: EdgeInsets.only(top: 10), child: NotifAcara()),
+            Padding(padding: EdgeInsets.only(top: 10), child: NotifAcara()),
+            Padding(padding: EdgeInsets.only(top: 10), child: NotifAcara()),
+          ],
+        ),
+      ),
+    );
+  }
+}
